scripts/requirements.txt
numpy==1.26.4
PyYAML==6.0.1
matplotlib==3.8.4
scikit-learn==1.4.1.post1
